move routes under devices
