+++ ui/user/src/routes/admin/device-clients/+page.ts
@@ -1,6 +1,8 @@
 import type { PageLoad } from './$types';
 import { redirect } from '@sveltejs/kit';
 
-export const load: PageLoad = async () => {
-	throw redirect(301, '/admin/devices?view=device-clients');
+export const load: PageLoad = ({ url }) => {
+	const searchParams = new URLSearchParams(url.searchParams);
+	searchParams.set('view', 'device-clients');
+	throw redirect(301, `/admin/devices?${searchParams}`);
 };
