+++ ui/user/src/routes/admin/device-skills/+page.ts
@@ -1,6 +1,8 @@
 import type { PageLoad } from './$types';
 import { redirect } from '@sveltejs/kit';
 
-export const load: PageLoad = async () => {
-	throw redirect(301, '/admin/devices?view=device-skills');
+export const load: PageLoad = ({ url }) => {
+	const searchParams = new URLSearchParams(url.searchParams);
+	searchParams.set('view', 'device-skills');
+	throw redirect(301, `/admin/devices?${searchParams}`);
 };
