+++ ui/user/src/routes/admin/devices/clients/+page.ts
@@ -1,0 +1,6 @@
+import type { PageLoad } from './$types';
+import { redirect } from '@sveltejs/kit';
+
+export const load: PageLoad = async () => {
+	throw redirect(301, '/admin/devices?view=device-clients');
+};
