+++ ui/user/src/routes/admin/devices/clients/[name]/+page.ts
@@ -1,0 +1,23 @@
+import { handleRouteError } from '$lib/errors';
+import {
+	AdminService,
+	UserService,
+	type DeviceClientFleetSummary,
+	type OrgUser
+} from '$lib/services';
+import { profile } from '$lib/stores';
+import type { PageLoad } from './$types';
+
+export const load: PageLoad = async ({ params, fetch }) => {
+	let client: DeviceClientFleetSummary;
+	let users: OrgUser[];
+	try {
+		[client, users] = await Promise.all([
+			AdminService.getDeviceClient(params.name, { fetch }),
+			UserService.listUsers({ fetch }).catch(() => [] as OrgUser[])
+		]);
+		return { client, users };
+	} catch (err) {
+		handleRouteError(err, `/admin/devices/clients/${params.name}`, profile.current);
+	}
+};
