+++ ui/user/src/routes/admin/devices/mcp-servers/[hash]/+page.svelte
