+++ ui/user/src/routes/admin/devices/mcp-servers/[hash]/+page.ts
@@ -1,0 +1,37 @@
+import { handleRouteError } from '$lib/errors';
+import { AdminService } from '$lib/services';
+import type {
+	DeviceMCPServerOccurrenceResponse,
+	DeviceMCPServerDetail
+} from '$lib/services/admin/types';
+import { profile } from '$lib/stores';
+import type { PageLoad } from './$types';
+
+const PAGE_SIZE = 50;
+
+export const load: PageLoad = async ({
+	params,
+	url,
+	fetch
+}: {
+	params: { hash: string };
+	url: URL;
+	fetch: typeof globalThis.fetch;
+}) => {
+	const offset = parseInt(url.searchParams.get('offset') ?? '0', 10) || 0;
+	let detail: DeviceMCPServerDetail | null;
+	let occurrences: DeviceMCPServerOccurrenceResponse;
+	try {
+		[detail, occurrences] = await Promise.all([
+			AdminService.getDeviceMCPServerDetail(params.hash, { fetch }),
+			AdminService.listDeviceMCPServerOccurrences(
+				params.hash,
+				{ limit: PAGE_SIZE, offset },
+				{ fetch }
+			)
+		]);
+		return { detail, occurrences, pageSize: PAGE_SIZE };
+	} catch (err) {
+		handleRouteError(err, `/admin/devices/mcp-servers/${params.hash}`, profile.current);
+	}
+};
