+++ ui/user/src/routes/admin/devices/skills/[name]/+page.svelte
